Transcribe Audio AI/requirements.txt
openai-whisper
torch
torchaudio
ffmpeg-python
pyannote.audio
speechbrain
